// packages/@aws-cdk/aws-cognito/lib/private/attr-names.d.ts
export declare const StandardAttributeNames: {
    address: string;
    birthdate: string;
    email: string;
    familyName: string;
    gender: string;
    givenName: string;
    locale: string;
    middleName: string;
    fullname: string;
    nickname: string;
    phoneNumber: string;
    profilePicture: string;
    preferredUsername: string;
    profilePage: string;
    timezone: string;
    lastUpdateTime: string;
    website: string;
    /** @deprecated */
    emailVerified: string;
    /** @deprecated */
    phoneNumberVerified: string;
};